Implemented four new control group classes & minor code style cleanup.
