+++ freeplane/src/main/java/org/freeplane/features/styles/mindmapmode/styleeditorpanel/EdgeDashControlGroup.java
@@ -19,64 +19,69 @@
  */
 package org.freeplane.features.styles.mindmapmode.styleeditorpanel;
 
-import java.awt.Color;
 import java.beans.PropertyChangeEvent;
 import java.util.List;
 
 import org.freeplane.core.resources.components.BooleanProperty;
-import org.freeplane.core.resources.components.ColorProperty;
+import org.freeplane.core.resources.components.ComboProperty;
 import org.freeplane.core.resources.components.IPropertyControl;
-import org.freeplane.core.util.ColorUtils;
+import org.freeplane.features.DashVariant;
 import org.freeplane.features.edge.EdgeController;
 import org.freeplane.features.edge.EdgeModel;
 import org.freeplane.features.edge.mindmapmode.MEdgeController;
 import org.freeplane.features.map.NodeModel;
+import org.freeplane.features.mode.Controller;
 
 /**
  * @author Joe Berry
- * Nov 27, 2016
+ * Dec 1, 2016
  */
-class EdgeColorControlGroup implements ControlGroup {
-	static final String EDGE_COLOR = "edgecolor";
+public class EdgeDashControlGroup implements ControlGroup {
+	private static final String EDGE_DASH = "edgedash";
 
-	private BooleanProperty mSetEdgeColor;
-	private ColorProperty mEdgeColor;
-	private EdgeColorChangeListener propertyChangeListener;
-
-	private class EdgeColorChangeListener extends ControlGroupChangeListener {
-		public EdgeColorChangeListener(final BooleanProperty mSet, final IPropertyControl mProperty) {
+	private BooleanProperty mSetEdgeDash;
+	private ComboProperty mEdgeDash;
+	private EdgeDashChangeListener propertyChangeListener;
+	
+	private class EdgeDashChangeListener extends ControlGroupChangeListener {
+		public EdgeDashChangeListener(final BooleanProperty mSet, final IPropertyControl mProperty) {
 			super(mSet, mProperty);
 		}
 
 		@Override
 		void applyValue(final boolean enabled, final NodeModel node, final PropertyChangeEvent evt) {
-			final MEdgeController edgeController = (MEdgeController) MEdgeController.getController();
-			edgeController.setColor(node, enabled ? mEdgeColor.getColorValue() : null);
+			final MEdgeController styleController = (MEdgeController) Controller
+			.getCurrentModeController().getExtension(
+					EdgeController.class);
+			styleController.setDash(node, enabled ? DashVariant.valueOf(mEdgeDash.getValue()): null);
 		}
 
 		@Override
 		void setStyleOnExternalChange(NodeModel node) {
-			final EdgeModel edgeModel = EdgeModel.getModel(node);
 			final EdgeController edgeController = EdgeController.getController();
+			final EdgeModel edgeModel = EdgeModel.getModel(node);
 			{
-				final Color edgeColor = edgeModel != null ? edgeModel.getColor() : null;
-				final Color viewColor = edgeController.getColor(node);
-				mSetEdgeColor.setValue(edgeColor != null);
-				mEdgeColor.setColorValue(viewColor);
+				final DashVariant dash = edgeModel != null ? edgeModel.getDash() : null;
+				final DashVariant viewDash = edgeController.getDash(node);
+				mSetEdgeDash.setValue(dash != null);
+				mEdgeDash.setValue(viewDash.name());
 			}
 		}
 	}
-	public void addControlGroup(final List<IPropertyControl> controls) {
-		mSetEdgeColor = new BooleanProperty(ControlGroup.SET_RESOURCE);
-		controls.add(mSetEdgeColor);
-		mEdgeColor = new ColorProperty(EdgeColorControlGroup.EDGE_COLOR, ColorUtils.colorToString(EdgeController.STANDARD_EDGE_COLOR));
-		controls.add(mEdgeColor);
-		propertyChangeListener = new EdgeColorChangeListener(mSetEdgeColor, mEdgeColor);
-		mSetEdgeColor.addPropertyChangeListener(propertyChangeListener);
-		mEdgeColor.addPropertyChangeListener(propertyChangeListener);
-	}
-	
+
+	@Override
 	public void setStyle(NodeModel node) {
 		propertyChangeListener.setStyle(node);
+	}
+
+	@Override
+	public void addControlGroup(List<IPropertyControl> controls) {
+		mSetEdgeDash = new BooleanProperty(ControlGroup.SET_RESOURCE);
+		controls.add(mSetEdgeDash);
+		mEdgeDash = ComboProperty.of(EDGE_DASH, DashVariant.class);
+		controls.add(mEdgeDash);
+		propertyChangeListener = new EdgeDashChangeListener(mSetEdgeDash, mEdgeDash);
+		mSetEdgeDash.addPropertyChangeListener(propertyChangeListener);
+		mEdgeDash.addPropertyChangeListener(propertyChangeListener);
 	}
 }
